* r504 release sp build-21255

+++ client/QueueManager.cpp
@@ -586,7 +586,7 @@ size_t QueueManager::FileQueue::getRunningFileCount(const size_t p_stop_key)
 	return l_cnt;
 }
 
-void QueueManager::FileQueue::calcPriorityAndGetRunningFilesL(QueueItem::PriorityArray& p_changedPriority, QueueItemList& p_runningFiles)
+void QueueManager::FileQueue::calcPriorityAndGetRunningFilesL(bool p_is_calc_prior, QueueItem::PriorityArray& p_changedPriority, QueueItemList& p_runningFiles)
 {
 	for (auto i = g_queue.cbegin(); i != g_queue.cend(); ++i)
 	{
@@ -596,20 +596,25 @@ void QueueManager::FileQueue::calcPriorityAndGetRunningFilesL(QueueItem::Priorit
 		if (q->isRunning())
 		{
 			p_runningFiles.push_back(q);
-			// TODO calcAverageSpeedAndDownloadedBytes - тяжелая операция зачем зовем так часто?
-			q->calcAverageSpeedAndCalcAndGetDownloadedBytesL();
-			if (q->getAutoPriority())
-			{
-				const QueueItem::Priority p1 = q->getPriority();
-				if (p1 != QueueItem::PAUSED)
-				{
-					const QueueItem::Priority p2 = q->calculateAutoPriority();
-					if (p1 != p2)
-					{
-						p_changedPriority.push_back(make_pair(q->getTarget(), p2));
-					}
-				}
-			}
+#ifdef FLYLINKDC_USE_RECALC_PRIOR	
+            if (p_is_calc_prior)
+            {
+                // TODO calcAverageSpeedAndDownloadedBytes - тяжелая операция зачем зовем так часто?
+                q->calcAverageSpeedAndCalcAndGetDownloadedBytesL();
+                if (q->getAutoPriority())
+                {
+                    const QueueItem::Priority p1 = q->getPriority();
+                    if (p1 != QueueItem::PAUSED)
+                    {
+                        const QueueItem::Priority p2 = q->calculateAutoPriority();
+                        if (p1 != p2)
+                        {
+                            p_changedPriority.push_back(make_pair(q->getTarget(), p2));
+                        }
+                    }
+                }
+            }
+#endif // FLYLINKDC_USE_RECALC_PRIOR
 		}
 	}
 }
@@ -908,6 +913,7 @@ QueueManager::~QueueManager() noexcept
 	rechecker.waitShutdown();
 	// [~] IRainman core.
 	saveQueue();
+#ifdef FLYLINKDC_USE_KEEP_LISTS
 	
 	if (!BOOLSETTING(KEEP_LISTS))
 	{
@@ -921,6 +927,7 @@ QueueManager::~QueueManager() noexcept
 			              File::deleteFile);
 		}
 	}
+#endif
 	SharedFileStream::check_before_destoy();
 }
 
@@ -2572,9 +2579,12 @@ bool QueueManager::removeTarget(const string& aTarget, bool p_is_batch_remove)
 		}
 		else if (!l_temp_target.empty() && l_temp_target != q->getTarget())
 		{
-			if (!File::deleteFile(l_temp_target))
+			if (File::isExist(l_temp_target))
 			{
-				SharedFileStream::delete_file(l_temp_target);
+				if (!File::deleteFile(l_temp_target))
+				{
+					SharedFileStream::delete_file(l_temp_target);
+				}
 			}
 		}
 		
@@ -2731,7 +2741,13 @@ void QueueManager::setPriority(const string& aTarget, QueueItem::Priority p) noe
 					// Problem, we have to request connections to all these users...
 					q->getOnlineUsers(l_getConn);
 				}
-				g_userQueue.setQIPriority(q, p); // !!!!!!!!!!!!!!!!!! Удаляет и вставляет в массив каждую секунду
+                // тут возникает проблема https://github.com/pavel-pimenov/flylinkdc-r5xx/issues/1692
+				//g_userQueue.setQIPriority(q, p); // !!!!!!!!!!!!!!!!!! Удаляет и вставляет в массив каждую секунду
+                // Поменял вызов на 
+                q->setPriority(p);
+                // Ошибка активировалась между билдами 21194 и 21203
+                // в этой ревизии я чинил баг - Убрал прерывание закачки большого кол-во файлов (revert r20612)
+                // В общем путное место.
 				
 #ifdef _DEBUG
 				LogManager::message("QueueManager g_userQueue.setQIPriority q->getTarget = " + q->getTarget());
@@ -3035,6 +3051,7 @@ void QueueLoader::endTag(const string& name, const string&)
 		}
 	}
 }
+#ifdef FLYLINKDC_USE_KEEP_LISTS
 
 void QueueManager::noDeleteFileList(const string& path)
 {
@@ -3043,6 +3060,7 @@ void QueueManager::noDeleteFileList(const string& path)
 		protectedFileLists.push_back(path);
 	}
 }
+#endif
 
 // SearchManagerListener
 void QueueManager::on(SearchManagerListener::SR, const std::unique_ptr<SearchResult>& p_sr) noexcept
@@ -3222,7 +3240,7 @@ void QueueManager::on(TimerManagerListener::Second, uint64_t aTick) noexcept
 		if ((g_filter++ % 10) == 0) // Делаем расчет приоритетов реже
 		{
 			RLock(*FileQueue::g_csFQ);
-			calcPriorityAndGetRunningFilesL(l_priorities, l_runningItems);
+			calcPriorityAndGetRunningFilesL(false, l_priorities, l_runningItems);
 		}
 	}
 	if (!l_runningItems.empty())
@@ -3245,11 +3263,12 @@ void QueueManager::on(TimerManagerListener::Second, uint64_t aTick) noexcept
 			fly_fire1(QueueManagerListener::TargetsUpdated(), l_fire_src_array);
 		}
 	}
-	
+#ifdef FLYLINKDC_USE_RECALC_PRIOR	
 	for (auto p = l_priorities.cbegin(); p != l_priorities.cend(); ++p)
 	{
 		setPriority(p->first, p->second);
 	}
+#endif
 }
 
 #ifdef FLYLINKDC_USE_DROP_SLOW
